# Conflicts:
#	src/main/java/com/golmok/golmokstar/controller/MapPinController.java
#	src/main/java/com/golmok/golmokstar/controller/PlaceController.java
#	src/main/java/com/golmok/golmokstar/controller/TripController.java
#	src/main/java/com/golmok/golmokstar/controller/TripParticipantController.java
#	src/main/java/com/golmok/golmokstar/dto/MapPinRequestDto.java
#	src/main/java/com/golmok/golmokstar/dto/MapPinResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/MapPinVisitRequestDto.java
#	src/main/java/com/golmok/golmokstar/dto/PlaceDetailResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/PlaceListResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/PlaceRequestDto.java
#	src/main/java/com/golmok/golmokstar/dto/PlaceResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripCreateRequestDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripDeleteResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripDetailResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripListResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripParticipantListResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripParticipantRequestDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripParticipantResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripResponseDto.java
#	src/main/java/com/golmok/golmokstar/dto/TripUpdateRequestDto.java
#	src/main/java/com/golmok/golmokstar/entity/MapPin.java
#	src/main/java/com/golmok/golmokstar/entity/Place.java
#	src/main/java/com/golmok/golmokstar/entity/Trip.java
#	src/main/java/com/golmok/golmokstar/entity/TripParticipant.java
#	src/main/java/com/golmok/golmokstar/enums/PinType.java
#	src/main/java/com/golmok/golmokstar/exception/CustomException.java
#	src/main/java/com/golmok/golmokstar/exception/GlobalExceptionHandler.java
#	src/main/java/com/golmok/golmokstar/repository/MapPinRepository.java
#	src/main/java/com/golmok/golmokstar/repository/PlaceRepository.java
#	src/main/java/com/golmok/golmokstar/repository/TripParticipantRepository.java
#	src/main/java/com/golmok/golmokstar/repository/TripRepository.java
#	src/main/java/com/golmok/golmokstar/service/MapPinService.java
#	src/main/java/com/golmok/golmokstar/service/PlaceService.java
#	src/main/java/com/golmok/golmokstar/service/TripParticipantService.java
#	src/main/java/com/golmok/golmokstar/service/TripService.java

+++ src/main/java/com/golmok/golmokstar/entity/MapPin.java
@@ -16,7 +16,7 @@ public class MapPin {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer pinId;
+    private Long pinId;  // ✅ Integer → Long으로 변경
 
     @ManyToOne
     @JoinColumn(name = "trip_id", nullable = false)
